requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
reportlab==4.0.7
python-dotenv==1.0.0
pytz==2023.3
